Explicitly use msgpack from homebrew
When homebrew is installed in a non-default path (e.g. under `~`)
this formula wasn't able to find the msgpack dylib.

This change explicitly configures the build to use the msgpack dylib
from the path in which homebrew installed it.

This should fix errors encountered at install like
`error on install: dyld: Library not loaded: @rpath/libmsgpackc.2.dylib`